0/0.8/0.8.3/requirements.txt
pinecone>=4.0.0
pandas>=1.5.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
streamlit>=1.22.0
plotly>=5.13.0
numpy>=1.24.0
sentence-transformers>=2.2.0